WORKING AUTONOMOUS CODE
this follows the error in QF4, where autonomous indeed caused the robot to fail. for a similar error that was previously investigated. and solved. we just need to recompile the code before matches.
watch the "differential drive not updated often enough error" because that's the problem

so.... what i did to fix this was add a print statement at line 112, compile and run. worked like a charm. then i removed the print statement, compiled, and ran. worked like a charm. no difference between the new and the original except recompiling.

it might have just been an internal roboRIO error but this was very highly oddly unusual.

will discuss in-depth on tuesday (tomorrow)