IMPORTANT:
Current code compiles but will not deploy given that motors are assigned twice. We need a screw class, and then we will be able to delete the ball class and that will make this current set of code deployable. To work on limelight, I intend to temporirly revert back to previous code