Fixed limelight LED code and started on second autonomous run
Two issues: First, the challenge states that we must RANDOMLY determine the path we will run and cannot communicate that chosen path to the robot. We need some way for the robot to determine for itself which path it should take. Second, the robot seems to be drifting while driving straight. (This is an issue with the drive, NOT the gyroscope, as the problem was there during Teleop)