Various Unimportant Clean-Up
Robot.java - Condensed teleopInit() using the setStatus() helper method. Added "Auto" to available statuses and applied to autonomousInit(). Limelight now toggles on one switch instead of two. Limelight aligning code fixed.
Limelight.java - New getter method for if the limelight is on or not. Limelight aligning code fixed.
Shooter.java - Ensured that killing the shooter will turn off the motor in the event of a different mode. Should not ever occur, but just in case.